% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{build_request}
\alias{build_request}
\title{Build a CDS API Request}
\usage{
build_request(
  model,
  variable,
  scenario,
  start_year,
  end_year,
  months,
  region,
  temporal_resolution = "daily"
)
}
\arguments{
\item{model}{Character. CMIP6 model name.}

\item{variable}{Character. Climate variable.}

\item{scenario}{Character. Emissions scenario.}

\item{start_year}{Integer. Start year.}

\item{end_year}{Integer. End year.}

\item{months}{Integer vector. Months to request.}

\item{region}{List. Output of \code{resolve_region()}.}
}
\value{
A list formatted for the CDS API.
}
\description{
Build a CDS API Request
}
\keyword{internal}
