% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cmip6_dir.R
\name{get_cmip6_dir}
\alias{get_cmip6_dir}
\title{Get the current CMIP6 data directory}
\usage{
get_cmip6_dir()
}
\value{
Character. Path to the current data directory.
}
\description{
Get the current CMIP6 data directory
}
\keyword{internal}
