% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip_dir.R
\name{set_cmip6_dir}
\alias{set_cmip6_dir}
\title{Set the directory for CMIP6 data downloads}
\usage{
set_cmip6_dir(path)
}
\arguments{
\item{path}{Character. Full path to the directory where downloaded \code{.nc}
files will be saved. The directory will be created if it does not exist.}
}
\description{
Set the directory for CMIP6 data downloads
}
\examples{
\dontrun{
set_cmip6_dir("C:/Users/nicol/Desktop/EAGLE/data")
}

}
