% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cmip6.R
\name{theme_cmip6}
\alias{theme_cmip6}
\title{ggplot2 Theme for cmip6r Plots}
\usage{
theme_cmip6(base_size = 13, base_family = "lora")
}
\arguments{
\item{base_size}{Numeric. Base font size. Default: \code{13}.}

\item{base_family}{Character. Font family. Default: \code{"lora"}.}
}
\value{
A ggplot2 theme object.
}
\description{
A clean minimal ggplot2 theme used as default for all cmip6r plots.
Automatically loads the Lora Google Font via \code{showtext}.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_cmip6()
}

}
