% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model.R
\name{validate_model}
\alias{validate_model}
\title{Validate CMIP6 Model Name}
\usage{
validate_model(model)
}
\arguments{
\item{model}{Character. Model name to validate.}
}
\value{
Invisibly returns \code{TRUE} if valid, otherwise throws an error.
}
\description{
Validate CMIP6 Model Name
}
\keyword{internal}
