% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cds_data.R
\name{load_cds_data}
\alias{load_cds_data}
\title{Download Data from the CDS API}
\usage{
load_cds_data(request, output_file)
}
\arguments{
\item{request}{List. A CDS-formatted request, output of \code{build_request()}.}

\item{output_file}{Character. Path to save the downloaded \code{.nc} file.}
}
\value{
Character. Path to the downloaded file.
}
\description{
Download Data from the CDS API
}
\keyword{Download}
\keyword{ZIP}
\keyword{dem}
\keyword{entpacken}
\keyword{internal}
\keyword{load_cds_data.R}
\keyword{nach}
\keyword{–}
