% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_time.R
\name{resolve_time}
\alias{resolve_time}
\title{Resolve Time Range for a CMIP6 Scenario}
\usage{
resolve_time(scenario, start_year = NULL, end_year = NULL)
}
\arguments{
\item{scenario}{Character. One of \code{"historical"}, \code{"ssp126"}, \code{"ssp245"},
\code{"ssp370"}, \code{"ssp585"}.}

\item{start_year}{Integer or NULL. Start year. Defaults to scenario minimum.}

\item{end_year}{Integer or NULL. End year. Defaults to scenario maximum.}
}
\value{
A list with \code{start_year} and \code{end_year}.
}
\description{
Resolve Time Range for a CMIP6 Scenario
}
\keyword{internal}
